♻️refactored output module
